@@ -1,10 +1,16 @@
 module golang.zx2c4.com/wireguard
 
-go 1.18
+go 1.23.1
 
 require (
-	golang.org/x/crypto v0.0.0-20220315160706-3147a52a75dd
-	golang.org/x/net v0.0.0-20220225172249-27dd8689420f
-	golang.org/x/sys v0.0.0-20220315194320-039c03cc5b86
-	golang.zx2c4.com/wintun v0.0.0-20211104114900-415007cec224
+	golang.org/x/crypto v0.37.0
+	golang.org/x/net v0.39.0
+	golang.org/x/sys v0.32.0
+	golang.zx2c4.com/wintun v0.0.0-20230126152724-0fa3db229ce2
+	gvisor.dev/gvisor v0.0.0-20250503011706-39ed1f5ac29c
+)
+
+require (
+	github.com/google/btree v1.1.2 // indirect
+	golang.org/x/time v0.7.0 // indirect
 )
