@@ -4,7 +4,6 @@
 // license that can be found in the LICENSE file.
 
 //go:build windows
-// +build windows
 
 // Package namedpipe implements a net.Conn and net.Listener around Windows named pipes.
 package namedpipe
@@ -29,7 +28,7 @@ type pipe struct {
 
 type messageBytePipe struct {
 	pipe
-	writeClosed int32
+	writeClosed atomic.Bool
 	readEOF     bool
 }
 
@@ -51,17 +50,17 @@ func (f *pipe) SetDeadline(t time.Time) error {
 
 // CloseWrite closes the write side of a message pipe in byte mode.
 func (f *messageBytePipe) CloseWrite() error {
-	if !atomic.CompareAndSwapInt32(&f.writeClosed, 0, 1) {
+	if !f.writeClosed.CompareAndSwap(false, true) {
 		return io.ErrClosedPipe
 	}
 	err := f.file.Flush()
 	if err != nil {
-		atomic.StoreInt32(&f.writeClosed, 0)
+		f.writeClosed.Store(false)
 		return err
 	}
 	_, err = f.file.Write(nil)
 	if err != nil {
-		atomic.StoreInt32(&f.writeClosed, 0)
+		f.writeClosed.Store(false)
 		return err
 	}
 	return nil
@@ -70,7 +69,7 @@ func (f *messageBytePipe) CloseWrite() error {
 // Write writes bytes to a message pipe in byte mode. Zero-byte writes are ignored, since
 // they are used to implement CloseWrite.
 func (f *messageBytePipe) Write(b []byte) (int, error) {
-	if atomic.LoadInt32(&f.writeClosed) != 0 {
+	if f.writeClosed.Load() {
 		return 0, io.ErrClosedPipe
 	}
 	if len(b) == 0 {
