@@ -1,3 +1,3 @@
 package main
 
-const Version = "0.0.20220316"
+const Version = "0.0.20250522"
